requirements.txt
djangorestframework==3.11.0
django==3.1.3
django-rest-auth==0.9.5
